+++ packages/openfermion/tests/openfermion/utils/test_add_exp_excitation_gates_trotter_decomposition.py
@@ -24,7 +24,7 @@ from quri_parts.openfermion.utils.add_exp_excitation_gates_trotter_decomposition
 
 class TestCreateAntiHermitianSdExcitationOperator:
     def test_single_excitation_jw(self) -> None:
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals=4)
 
         s_exc = (0, 2)
         op = create_anti_hermitian_sd_excitation_operator(s_exc, jw_mapper)
@@ -41,7 +41,7 @@ class TestCreateAntiHermitianSdExcitationOperator:
         assert op == expected_op
 
     def test_single_excitation_bk(self) -> None:
-        bk_mapper = bravyi_kitaev.get_of_operator_mapper()
+        bk_mapper = bravyi_kitaev.get_of_operator_mapper(n_spin_orbitals=4)
 
         s_exc = (0, 2)
         op = create_anti_hermitian_sd_excitation_operator(s_exc, bk_mapper)
@@ -75,7 +75,7 @@ class TestCreateAntiHermitianSdExcitationOperator:
         assert op == expected_op
 
     def test_double_excitation_jw(self) -> None:
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals=4)
 
         d_exc = (0, 1, 2, 3)
         op = create_anti_hermitian_sd_excitation_operator(d_exc, jw_mapper)
@@ -110,7 +110,7 @@ class TestCreateAntiHermitianSdExcitationOperator:
         assert op == expected_op
 
     def test_double_excitation_bk(self) -> None:
-        bk_mapper = bravyi_kitaev.get_of_operator_mapper()
+        bk_mapper = bravyi_kitaev.get_of_operator_mapper(n_spin_orbitals=8)
 
         d_exc = (0, 1, 2, 3)
         op = create_anti_hermitian_sd_excitation_operator(d_exc, bk_mapper)
@@ -183,7 +183,7 @@ class TestAddExpExcitationGatesTrotterDecomposition:
     def test_trotter1_jw(self) -> None:
         n_spin_orbitals = 4
         excitation_indices = [(0, 2)]
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals)
         trotter_number = 1
 
         circuit = LinearMappedUnboundParametricQuantumCircuit(n_spin_orbitals)
@@ -209,7 +209,7 @@ class TestAddExpExcitationGatesTrotterDecomposition:
     def test_trotter2_jw(self) -> None:
         n_spin_orbitals = 4
         excitation_indices = [(0, 2)]
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals)
         trotter_number = 2
 
         circuit = LinearMappedUnboundParametricQuantumCircuit(n_spin_orbitals)
@@ -235,7 +235,7 @@ class TestAddExpExcitationGatesTrotterDecomposition:
     def test_trotter1_bk(self) -> None:
         n_spin_orbitals = 4
         excitation_indices = [(0, 2)]
-        bk_mapper = bravyi_kitaev.get_of_operator_mapper()
+        bk_mapper = bravyi_kitaev.get_of_operator_mapper(n_spin_orbitals)
         trotter_number = 1
 
         circuit = LinearMappedUnboundParametricQuantumCircuit(n_spin_orbitals)
@@ -285,7 +285,7 @@ class TestAddExpExcitationGatesTrotterDecomposition:
     def test_doubles(self) -> None:
         n_spin_orbitals = 6
         excitation_indices = [(0, 1, 4, 5)]
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals)
         trotter_number = 1
 
         circuit = LinearMappedUnboundParametricQuantumCircuit(n_spin_orbitals)
@@ -334,7 +334,7 @@ class TestAddExpPauliGatesFromLinearMappedFunctions:
 
     def test_jw(self) -> None:
         n_spin_orbitals = 6
-        jw_mapper = jordan_wigner.get_of_operator_mapper()
+        jw_mapper = jordan_wigner.get_of_operator_mapper(n_spin_orbitals)
 
         circuit = LinearMappedUnboundParametricQuantumCircuit(n_spin_orbitals)
         param1 = circuit.add_parameter("param1")
